rename extract_and_sell behaviour
