+++ .vscode/launch.json
@@ -13,7 +13,7 @@
             "name": "conductor",
             "type": "python",
             "request": "launch",
-            "program": "conductorWK11.py",
+            "program": "conductorWK12.py",
             "console": "integratedTerminal",
             "justMyCode": false
         },
@@ -21,7 +21,7 @@
             "name": "dispatcher -U",
             "type": "python",
             "request": "launch",
-            "program": "dispatcherWK11.py",
+            "program": "dispatcherWK12.py",
             "console": "integratedTerminal",
             "justMyCode": false,
             "args": [
