+++ conductorWK12.py
@@ -25,7 +25,7 @@ import hashlib
 import logging
 import time
 from datetime import datetime, timedelta
-from dispatcherWK9 import (
+from dispatcherWK12 import (
     BHVR_EXTRACT_AND_SELL,
     BHVR_RECEIVE_AND_FULFILL,
     BHVR_EXTRACT_AND_TRANSFER_OR_SELL,
