+++ old_conductors_and_dispatchers/dispatcherWK7.py
@@ -14,7 +14,7 @@ from behaviours.receive_and_fulfill import ReceiveAndFulfillOrSell_3
 from behaviours.generic_behaviour import Behaviour
 import random
 from pyrate_limiter import Limiter, Duration, RequestRate
-from behaviours.extract_and_transfer_or_sell import (
+from behaviours.extract_and_transfer_or_sell_wk11 import (
     ExtractAndTransferOrSell_4,
     BEHAVIOUR_NAME as BHVR_EXTRACT_AND_TRANSFER_OR_SELL,
 )
